model: transformers: Move out of core plugins
The transformers pacakge is not compatable with the latest version of
tensorflow (2.4.x). We do not have time to update the transformers API
calls to the 3.x versions. We instead are temporarilly removing it from
teh set of core plugins maintained within the DFFML source repo.

dffml-model-transformers has been moved to
https://github.com/dffml/dffml-model-transformers

Related: #844

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>